features: Move feature-handling code to a common/features.c source.
